UPDATED: arcemu-world VC++ 100 project to latest AIInterface changes FIXED: Some warnings. MODIFIED: GetHealthPct() returns an unsigned int instead of a signed int. Note: right now only "arcemuVC100 - All Projects - no scripts.sln" builds.

+++ src/arcemu-world/AI_Guardian.cpp
@@ -279,7 +279,7 @@ Unit * GuardianAI::getSpellTarget(AI_Spell* nSpell)
 		/* If we are too far, we simply move to the max range area and try to cast. */
 		if(nSpell->minrange > 0 && dist < nSpell->minrange ||
 			nSpell->maxrange > 0 && dist > nSpell->maxrange ||
-			(uint32)target->GetHealthPct() > hspell->heal_hp )
+			target->GetHealthPct() > hspell->heal_hp )
 		{	// Check health requirements, if we fail the hp requirements, try again in 2secs.
 			if(target->GetHealthPct() > hspell->heal_hp)
 			{
